Repo Reorg: move verify* functions to top level
