--- src/nisarqa/products/__init__.py
@@ -1,2 +1,0 @@
-# keep individual products in their own namespace
-from . import gcov, gslc, igram, offsets, rslc
