+++ src/nisarqa/verify_offsets.py
